#9862 move image_scaling task to convert worker
move scaling_image from default worker to convert worker,
delete pillow dependency from default-worker's Pipfile
to reduce the number of dependencies
