+++ service/job/job_type_controller.py
@@ -182,7 +182,7 @@ def get_job_type_detail(job_type):
                         "foreach": {
                             "max_height": 50,
                             "max_width": 50,
-                            "task": "image.scaling"
+                            "task": "convert.image_scaling"
                         },
                         "list_files": "jpg"
                     },
@@ -193,7 +193,7 @@ def get_job_type_detail(job_type):
                         "foreach": {
                             "max_height": 50,
                             "max_width": 50,
-                            "task": "image.scaling"
+                            "task": "convert.image_scaling"
                         },
                         "list_files": "tif"
                     },
