+++ workers/default/tasks.py
@@ -4,6 +4,5 @@ celery_app.autodiscover_tasks([
     'workers.default.repository',
     'workers.default.utils',
     'workers.default.xml',
-    'workers.default.ojs',
-    'workers.default.image'
+    'workers.default.ojs'
     ], force=True)
